+++ apps/etterlatte-oppdater-behandling/src/test/kotlin/regulering/OmregningsHendelserBehandlingRiverTest.kt
@@ -6,7 +6,8 @@ import io.mockk.slot
 import no.nav.etterlatte.BehandlingServiceImpl
 import no.nav.etterlatte.libs.common.behandling.Omregningshendelse
 import no.nav.etterlatte.libs.common.behandling.SakType
-import no.nav.etterlatte.libs.common.omregning.OpprettOmregningResponse
+import no.nav.etterlatte.libs.common.revurdering.AutomatiskRevurderingRequest
+import no.nav.etterlatte.libs.common.revurdering.AutomatiskRevurderingResponse
 import no.nav.etterlatte.rapidsandrivers.BEHANDLING_ID_KEY
 import no.nav.helse.rapids_rivers.testsupport.TestRapid
 import org.junit.jupiter.api.Assertions
@@ -24,7 +25,7 @@ internal class OmregningsHendelserBehandlingRiverTest {
         val behandlingId = UUID.randomUUID()
         val behandlingViOmregnerFra = UUID.randomUUID()
 
-        val returnValue = OpprettOmregningResponse(behandlingId, behandlingViOmregnerFra, SakType.BARNEPENSJON)
+        val returnValue = AutomatiskRevurderingResponse(behandlingId, behandlingViOmregnerFra, SakType.BARNEPENSJON)
 
         every { behandlingService.opprettOmregning(capture(omregningshendelseSlot)) }.returns(returnValue)
